refactor: Modify the animation component, system
The animation component has unnecessary separation, so we combine it from three to two.

The animation system has multiple responsibilities in one system, so it separates from three to four.

+++ WITH_Server/WITH_Server/AnimationManager.h
@@ -53,7 +53,7 @@ struct StaticCapsuleData {
 
 struct PrebakedAnimation {
 	float fps;
-	uint8 numFrames;
+	uint16 numFrames;
 
 	std::vector<StaticCapsuleData> staticDatas;
 	std::vector<std::vector<DynamicCapsuleData>> dynamicDatas;
